server/requirements.txt
requests
python-dotenv
google-genai
pymongo
pytz
fastapi
uvicorn
typing-extensions
python-multipart
beautifulsoup4
lxml
aiohttp
selenium
webdriver-manager
selenium-stealth